refactor(datasets,projects): decompose createDataset action into primitives
Both `datasets create` and `projects create` now compose
`determineDatasetAclMode` + `createDataset` service directly instead
of going through the bundled action. Spinners and success logs live
in the commands.

Also fixes `projects create` not passing `forcePublic` when user
picks default config — visibility prompt was shown regardless.
